WIP apply default types for training plan torch

+++ fedbiomed/common/dataset/_tabular_dataset.py
@@ -18,7 +18,7 @@ class TabularDataset(Dataset):
 
     _native_to_framework = {
         DataReturnFormat.SKLEARN: lambda x: x.to_numpy(),
-        DataReturnFormat.TORCH: lambda x: x.to_torch().float(),
+        DataReturnFormat.TORCH: lambda x: x.to_torch(),
     }
 
     def __init__(
